// src/LineChart/models/sizes.ts
export type LineChartSizes = {
    left?: number;
    bottom?: number;
    right?: number;
    top?: number;
    width?: number;
    height?: number;
}


export interface ChartSizesInfo {
    left: number;
    bottom: number;
    right: number;
    top: number;
    width: number;
    height: number;
    innerHeight: number;
    innerWidth: number;
}